simulated_user_testing/requirements.txt
awscli>=1.18.152,<2.0
pycrunch>=0.4.13,<1.0
docopt>=0.6.2,<1.0
honeycomb-beeline>=2.12.1,<3.0
libhoney>=1.9.0,<2.0
ddtrace>=0.43.0,<1.0
six>=1.15.0,<2.0
pyyaml>=3.12,<4.0
ipython
